src/requirements.txt
boto==2.39.0
braintree==3.24.0
dj-database-url==0.3.0
dj-stripe==0.8.0
Django==1.8
django-braces==1.8.1
django-crispy-forms==1.5.2
django-filter==0.12.0
django-model-utils==2.4
django-registration-redux==1.2
django-storages==1.1.8
gunicorn==19.4.5
httpie==0.9.3
jsonfield==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
Pygments==2.1
pytz==2015.7
requests==2.9.1
six==1.10.0
stripe==1.29.1
whitenoise==2.0.6
